requirements/base.txt
Django==3.0.6
celery==4.4.2
django-widget-tweaks==1.4.8
psycopg2-binary==2.8.5
openpyxl==3.0.3
docxtpl==0.9.2
docx==0.2.4
babel==2.7.0
beautifulsoup4==4.8.1
chardet==3.0.4
fake-useragent==0.1.11
requests==2.22.0
pylibmc==1.6.1
pytelegrambotapi==3.7.1
xmltodict==0.12.0
xlrd==1.2.0
django-formtools==2.2
django-extensions==2.2.9
matplotlib==3.2.1
djangorestframework==3.11.0
django_filter==2.2.0
django-cors-headers==3.2.1
flower==0.9.4
django-environ==0.4.5
